requirements.txt
# core requirements
attrs==18.2.0
pyyaml==3.12

# extras for development
nose
nose-exclude
coverage
pylint==1.8.2
flake8==3.5.0
mypy==0.620
